requirements.txt
# Core dependencies
browser-use>=0.1.40
langchain-ollama>=0.1.0
langchain>=0.2.0
langchain-community>=0.2.0
beautifulsoup4>=4.12.0
lxml>=5.1.0


# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector database
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Google Sheets
gspread>=6.0.0
oauth2client>=4.1.3
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# Telegram
python-telegram-bot>=20.7

# Scheduler
apscheduler>=3.10.4

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0
loguru>=0.7.2
tenacity>=8.2.3

# Data processing
pandas>=2.1.0
phonenumbers>=8.13.0
python-dateutil>=2.8.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.0

# AI/ML
autogluon>=1.0.0
